+++ BraintreeCore/src/test/java/com/braintreepayments/api/core/usecase/GetReturnLinkUseCaseUnitTest.kt
@@ -1,6 +1,7 @@
-package com.braintreepayments.api.core
+package com.braintreepayments.api.core.usecase
 
 import android.net.Uri
+import com.braintreepayments.api.core.MerchantRepository
 import io.mockk.every
 import io.mockk.mockk
 import org.junit.Before
@@ -15,24 +16,31 @@ class GetReturnLinkUseCaseUnitTest {
 
     private val merchantRepository: MerchantRepository = mockk(relaxed = true)
     private val getReturnLinkTypeUseCase: GetReturnLinkTypeUseCase = mockk(relaxed = true)
+    private val getDefaultAppUseCase: GetDefaultAppUseCase = mockk(relaxed = true)
+    private val getAppLinksCompatibleBrowserUseCase: GetAppLinksCompatibleBrowserUseCase = mockk(relaxed = true)
     private val appLinkReturnUri = Uri.parse("https://example.com")
     private val deepLinkFallbackUrlScheme = "com.braintreepayments.demo"
 
-    lateinit var subject: GetReturnLinkUseCase
+    lateinit var sut: GetReturnLinkUseCase
 
     @Before
     fun setUp() {
         every { merchantRepository.appLinkReturnUri } returns appLinkReturnUri
         every { merchantRepository.deepLinkFallbackUrlScheme } returns deepLinkFallbackUrlScheme
 
-        subject = GetReturnLinkUseCase(merchantRepository, getReturnLinkTypeUseCase)
+        sut = GetReturnLinkUseCase(
+            merchantRepository,
+            getDefaultAppUseCase,
+            getAppLinksCompatibleBrowserUseCase,
+            getReturnLinkTypeUseCase
+        )
     }
 
     @Test
     fun `when invoke is called and app link is available, AppLink is returned`() {
         every { getReturnLinkTypeUseCase() } returns GetReturnLinkTypeUseCase.ReturnLinkTypeResult.APP_LINK
 
-        val result = subject()
+        val result = sut()
 
         assertEquals(GetReturnLinkUseCase.ReturnLinkResult.AppLink(appLinkReturnUri), result)
     }
@@ -41,7 +49,7 @@ class GetReturnLinkUseCaseUnitTest {
     fun `when invoke is called and app link is not available, DeepLink is returned`() {
         every { getReturnLinkTypeUseCase() } returns GetReturnLinkTypeUseCase.ReturnLinkTypeResult.DEEP_LINK
 
-        val result = subject()
+        val result = sut()
 
         assertEquals(GetReturnLinkUseCase.ReturnLinkResult.DeepLink(deepLinkFallbackUrlScheme), result)
     }
@@ -51,12 +59,12 @@ class GetReturnLinkUseCaseUnitTest {
         every { getReturnLinkTypeUseCase() } returns GetReturnLinkTypeUseCase.ReturnLinkTypeResult.DEEP_LINK
         every { merchantRepository.deepLinkFallbackUrlScheme } returns null
 
-        val result = subject()
+        val result = sut()
 
         assertTrue(result is GetReturnLinkUseCase.ReturnLinkResult.Failure)
         assertEquals(
             "Deep Link fallback return url is null",
-            (result as GetReturnLinkUseCase.ReturnLinkResult.Failure).exception.message
+            result.exception.message
         )
     }
 
@@ -65,12 +73,12 @@ class GetReturnLinkUseCaseUnitTest {
         every { getReturnLinkTypeUseCase() } returns GetReturnLinkTypeUseCase.ReturnLinkTypeResult.APP_LINK
         every { merchantRepository.appLinkReturnUri } returns null
 
-        val result = subject()
+        val result = sut()
 
         assertTrue(result is GetReturnLinkUseCase.ReturnLinkResult.Failure)
         assertEquals(
             "App Link Return Uri is null",
-            (result as GetReturnLinkUseCase.ReturnLinkResult.Failure).exception.message
+            result.exception.message
         )
     }
 }